utils refactor
